requirements.txt
zhipuai 
langchain 
chromadb 
pypdf 
python-dotenv 
streamlit 
langchain-community 
